requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pandas==2.2.2
python-multipart==0.0.9
python-dotenv==1.0.1
google-genai==0.3.0
